# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM system module",
        "description": "Authorization, user registration.Adding an order to the system,viewing the list of orders,viewing an individual order.",
        "version": "3.0"
    },
    "paths": {
...
        "/api/order/add": {
            "post": {
                "tags": [
                    "Add_Order"
                ],
                "summary": "Adding an order",
                "operationId": "Adding an order",
                "parameters": [
                    {
                        "name": "order_number",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request(something wrong with URL or parameters)"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Logged in but access to requested area is forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity (validation failed)"
                    },
                    "500": {
                        "description": "General server error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}